First commit creating an app for storing sequence
This is a real dumb implementation of a seq storage layer for Ensembl projects. We have a basic implementation of a DBIx::Class ORM layer to a schema and then a Mojolicous API over the top. We do not support large DNA sequences just those associated with things like genes and transcripts and proteins.