Swith from seq_type to mol_type
In an attempt to avoid schema complications in a many to many, type is an attribute of a molecule and no longer sequence. This is inline with sha1 being the unique constraint in seq. Also renamed stable_id to id and brought in an import.t test to ensure loading the schema still works.