CORS is hard
Done some reading of https://www.html5rocks.com/static/images/cors_server_flowchart.png and it seems that this is not as simple as I thought. The reuqest has two parts.

The first is the OPTIONS and Access-Control-Request-Method and then there is the repsonse for methods, age and allow-headers. When it's a request then the response is expose-headers. We always set Access-Control-Allow-Origin.