vmcdigest and trunc512 support.
Implementing the trunc512 algorithm, which is a sha512 digest truncated
to 24 bytes (48 hex characters). This is the same basis for the VMC
digest code. It is now possible to switch between trunc512 and
vmcdigest through reformatting of each digest. Schemas have
had sha1, sha256 and sha512 removed from them in favour of the
trunc512 algorithm.

Also now support case-insensitive searching by lowercasing all
output from the algorithms and before searching.