Support Transfer-Encoding
Seems the missing part of the puzzle was the TE header. When set we move into Transfer-Encoding mode. If we detect the Accept-Encoding header we do the more nomal gzip encoding and set Vary to Accept-Encoding. More clients support the second mode but the first maintains compatability with the spec.