Fixing DBIx build option issues
Creating a DBIx sequence storage layer did not work because of an error in the builder method which failed to pass through an appropriate Schema object. This now happens. All tests run.

In addition populate-db-dicts.pl has been made executable and SQLite database URLs have been clarified to have 4 / when giving a path (2 for schema, one for the path start & one for the actual path)